fix: notifications

+++ src/notification/notification.gateway.ts
@@ -23,6 +23,7 @@ import {
 import { SubtaskConfirmService } from "../confirmation/subtask-confirmation.service";
 import { SubtaskService } from "../task/subtask.service";
 import { User } from "../user/user.schema";
+import { Subtask } from "src/task/schemas";
 
 @WebSocketGateway({
   cors: {
@@ -35,6 +36,7 @@ export class NotificationGateway
 {
   constructor(
     @InjectModel(User.name) private userModel: Model<User>,
+    @InjectModel(Subtask.name) private subtaskModel: Model<Subtask>,
     private readonly jwtService: JwtService,
     private readonly subtConfService: SubtaskConfirmService,
     private readonly subtaskService: SubtaskService,
@@ -81,6 +83,11 @@ export class NotificationGateway
     { subtaskId, receiverId }: { subtaskId: string; receiverId: string }
   ): Promise<void> {
     const userId = this.findUserIdByConnection(client.id);
+    const foundSubtask = await this.subtaskModel.findById(subtaskId);
+    if (!foundSubtask) {
+      throw new WsException("Subtask was deleted");
+    }
+
     await this.subtConfService.removeSubtaskConfirmation(subtaskId);
     await this.subtaskService.updateSubtaskIsConf(userId, subtaskId, true);
     const notification = await this.notificationService.create({
@@ -100,6 +107,11 @@ export class NotificationGateway
     { subtaskId, receiverId }: { subtaskId: string; receiverId: string }
   ): Promise<void> {
     const userId = this.findUserIdByConnection(client.id);
+    const foundSubtask = await this.subtaskModel.findById(subtaskId);
+    if (!foundSubtask) {
+      throw new WsException("Subtask was deleted");
+    }
+
     await this.subtConfService.removeSubtaskConfirmation(subtaskId);
     await this.subtaskService.updateSubtaskIsConf(userId, subtaskId, false);
     const notification = await this.notificationService.create({
